+++ src/components/AttackToolDetail.vue
@@ -118,4 +118,4 @@ const onNavigateCase = (cKey: string) => {
   />
 </template>
 
-<style src="./drawer-detail-shared.css" scoped></style>
+<style src="./drawer-detail-shared.css"></style>