+++ src/components/AvoidanceDetail.vue
@@ -100,4 +100,4 @@ const onNavigateRisk = (rKey: string) => {
   />
 </template>
 
-<style src="./drawer-detail-shared.css" scoped></style>
+<style src="./drawer-detail-shared.css"></style>