+++ src/components/RiskDetail.vue
@@ -119,4 +119,4 @@ const onNavigateCase = (cKey: string) => {
   />
 </template>
 
-<style src="./drawer-detail-shared.css" scoped></style>
+<style src="./drawer-detail-shared.css"></style>