+++ src/components/ThreatActorDetail.vue
@@ -106,4 +106,4 @@ const onNavigateCase = (cKey: string) => {
   />
 </template>
 
-<style src="./drawer-detail-shared.css" scoped></style>
+<style src="./drawer-detail-shared.css"></style>